backend/requirements.txt
Django
django-soft-delete
djangorestframework
siwe-auth-django